requirements.txt
requests==2.32.4
dnspython==2.7.0
